refactor(channex): extract booking revision import service
