+++ backend/functions/UnifiedMessaging/business/channexBookingRevisionImportService.test.js
@@ -1,0 +1,8 @@
+const ChannexBookingRevisionImportService =
+  require("../.shared/channelManagement/services/channexBookingRevisionImportService.js").default;
+
+describe("ChannexBookingRevisionImportService", () => {
+  test("is importable from the shared ChannelManagement boundary", () => {
+    expect(typeof ChannexBookingRevisionImportService).toBe("function");
+  });
+});
